Fixed AppData, added icons to CurrentWeather, useWeatherMock is deleted
