+++ Nimbus/src/composables/CurrentData.ts
@@ -1,12 +1,17 @@
-import { reactive, ref } from "vue";
+import { ref, computed } from "vue";
+import hourlyData from '../../ExampleJSON/JSONExampleHourly.json'
 
-const time = reactive ({
-  hours: 0,
-  seconds: 0,
-  milliseconds: 0
-})
+const currentWeather = hourlyData.timelines.hourly[0]
 
-const weather = ref('cloudy')
+const weatherCodeToLabel: Record<number, string> = {
+  1000: 'Clear',
+  1001: 'Cloudy',
+  1100: 'Mostly Clear',
+  1101: 'Partly Cloudy',
+  1102: 'Mostly Cloudy',
+}
 
-export function useWeather() {return weather}
-export function useTime() {return time}
+const weatherLabel = computed(() => weatherCodeToLabel[currentWeather.values.weatherCode] || 'Unknown')
+
+export function useWeatherLabel() { return weatherLabel }
+export function useCurrentWeather() { return currentWeather }
